requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
decorator==5.1.1
et-xmlfile==1.1.0
future==0.18.2
geocoder==1.38.1
idna==3.4
numpy==1.23.5
oauthlib==3.2.2
openai==0.25.0
openpyxl==3.0.10
pandas==1.5.2
pandas-stubs==1.5.2.221124
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
ratelim==0.1.6
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
tqdm==4.64.1
tweepy==4.12.1
types-pytz==2022.6.0.1
typing-extensions==4.4.0
urllib3==1.26.13
